+++ src/ByteDev.Configuration.Environment/IEnvironmentVariableProvider.cs
@@ -74,7 +74,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         char GetCharOrDefault(string name, char defaultValue = '\0');
@@ -95,7 +95,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         bool GetBooleanOrDefault(string name, bool defaultValue = false);
@@ -116,7 +116,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         byte GetByteOrDefault(string name, byte defaultValue = 0);
@@ -137,7 +137,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         short GetInt16OrDefault(string name, short defaultValue = 0);
@@ -158,7 +158,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         int GetInt32OrDefault(string name, int defaultValue = 0);
@@ -179,7 +179,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         long GetInt64OrDefault(string name, long defaultValue = 0);
@@ -200,7 +200,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         float GetSingleOrDefault(string name, float defaultValue = 0);
@@ -221,7 +221,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         double GetDoubleOrDefault(string name, double defaultValue = 0);
@@ -242,7 +242,7 @@ namespace ByteDev.Configuration.Environment
         /// then the <paramref name="defaultValue" /> will be returned.
         /// </summary>
         /// <param name="name">Name of environment variable.</param>
-        /// <param name="defaultValue">Value to return if the environment variable does not exist.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
         /// <returns>Environment variable's value.</returns>
         /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
         decimal GetDecimalOrDefault(string name, decimal defaultValue = 0);
@@ -280,5 +280,17 @@ namespace ByteDev.Configuration.Environment
         /// <exception cref="T:ByteDev.Configuration.Environment.EnvironmentVariableNotExistException">Environment variable does not exist.</exception>
         /// <exception cref="T:ByteDev.Configuration.Environment.UnexpectedEnvironmentVariableTypeException">Environment variable value is not a defined name or number value of the enum.</exception>
         TEnum GetEnum<TEnum>(string name) where TEnum : struct, Enum;
+
+        /// <summary>
+        /// Retrieve an environment variable as a Enum. The environment variable value can be stored as the
+        /// string (name) enum representation or the defined number value. If it does not exist or it's value cannot be cast 
+        /// then the <paramref name="defaultValue" /> will be returned.
+        /// </summary>
+        /// <typeparam name="TEnum">Type of enum.</typeparam>
+        /// <param name="name">Name of environment variable.</param>
+        /// <param name="defaultValue">Value to return if the environment variable does not exist or cannot be cast.</param>
+        /// <returns>Environment variable's value.</returns>
+        /// <exception cref="T:System.ArgumentException"><paramref name="name" /> was null or empty.</exception>
+        TEnum GetEnumOrDefault<TEnum>(string name, TEnum defaultValue) where TEnum : struct, Enum;
     }
 }